Moving div_wrapper function to config/styling.php file
