+++ functions.php
@@ -39,6 +39,7 @@ require_once( 'config/support.php' );
 require_once( 'config/breadcrumbs.php' );
 require_once( 'config/pagination.php' );
 require_once( 'config/woocommerce.php' );
+require_once( 'config/styling.php' );
 
 // Theme Recommended & Required Plugins
 require_once( 'config/required_plugins.php' );
@@ -55,31 +56,6 @@ function theme_queue_js()
     }
 }
 add_action('get_header', 'theme_queue_js');
-
-function div_wrapper($content)
-{
-    $supported_formats = ["youtube", "vimeo", "dailymotion"];
-
-    // Match any iframes
-    $pattern = '~<iframe.*</iframe>|<embed.*</embed>~';
-    preg_match_all($pattern, $content, $matches);
-
-    foreach ($matches[0] as $match) {
-        foreach ($supported_formats as $format) {
-            // Make sure the iframe/embed is of a supported format to avoid wrapping any non-video iframes
-            if (strpos($match, $format) !== false) {
-                // Wrap matched iframe with div
-                $wrappedframe = '<div class="video-wrapper"><div class="video-inner">' . $match . '</div></div>';
-
-                // Replace original iframe with new in content
-                $content = str_replace($match, $wrappedframe, $content);
-            }
-        }
-    }
-
-    return $content;
-}
-add_filter('the_content', 'div_wrapper');
 
 // Add Bootstrap Styling to Password Protected form
 function bootstrap_password_form() {
